tools/requirements.txt
# commented out requirements are anticipated but not needed yet.
coincurve==13.0.0
# cryptography==2.8
# eth-keys==0.3.1
google-auth==1.14.0
google-cloud-secret-manager==0.12.3
# jwcrypto==0.7
pysha3==1.0.2
# requests==2.23.0
